requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
whitenoise>=6.6,<7.0
django-cors-headers>=4.3,<5.0
djangorestframework-simplejwt>=5.3,<6.0
